+++ backend/controllers/auth_controller.js
@@ -2,6 +2,7 @@ const auth = require('express').Router()
 const db = require('../models')
 const bcrypt = require('bcrypt')
 const jwt = require('jsonwebtoken')
+require('dotenv').config()
 
 const { User } = db
